Move code to interact with user out of sypgp to a separate utility package
