+++ pkg/sypgp/sypgp_test.go
@@ -14,10 +14,10 @@ import (
 	"net/http/httptest"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/sylabs/singularity/internal/pkg/test"
+	"github.com/sylabs/singularity/internal/pkg/util/fs"
 	useragent "github.com/sylabs/singularity/pkg/util/user-agent"
 	"golang.org/x/crypto/openpgp"
 	"golang.org/x/crypto/openpgp/armor"
@@ -516,177 +516,134 @@ func TestPrintEntities(t *testing.T) {
 	}
 }
 
-func TestAskQuestion(t *testing.T) {
+const (
+	dummyPassPhrase  = "fakepassphrase"
+	validKeyGenInput = "A tester\ntest@my.info\n\n" + dummyPassPhrase + "\n" + dummyPassPhrase + "\nn\n"
+	myToken          = "MyToken"
+	myURI            = "MyURI"
+)
+
+func TestExportKeys(t *testing.T) {
 	test.DropPrivilege(t)
 	defer test.ResetPrivilege(t)
 
-	// Each line of the string represents a virtual different answer from a user
-	testStr := "test test test\ntest2 test2\n\ntest3"
-	testBytes := []byte(testStr)
-
-	// we create a temporary file that will act as Stdin
-	testFile, err := ioutil.TempFile("", "inputTest")
+	// Create a temporary keyring
+	tempDir, err := ioutil.TempDir("", "keyring-")
 	if err != nil {
-		t.Fatalf("failed to create temporary file: %s", err)
+		t.Fatalf("failed to create temporary keyring: %s", err)
 	}
-	defer testFile.Close()
-	defer os.Remove(testFile.Name())
-
-	// Write the data that AskQuestion() will later on need
-	_, err = testFile.Write(testBytes)
+	defer os.RemoveAll(tempDir)
+	keyring := NewHandle(tempDir)
+	secretDir := keyring.SecretPath()
+	publicDir := keyring.PublicPath()
+	err = fs.Touch(secretDir)
 	if err != nil {
-		t.Fatalf("failed to write to %s: %s", testFile.Name(), err)
+		t.Fatalf("failed to create file for secret keys: %s", err)
 	}
-
-	// Reposition to the beginning of file to ensure there is something to read
-	_, err = testFile.Seek(0, os.SEEK_SET)
+	err = fs.Touch(publicDir)
 	if err != nil {
-		t.Fatalf("failed to seek to beginning of file %s: %s", testFile.Name(), err)
+		t.Fatalf("failed to create file for public keys: %s", err)
 	}
 
-	// Redirect Stdin
+	privKeyExportFile := filepath.Join(tempDir, "priv-keys")
+	pubKeyExportFile := filepath.Join(tempDir, "pub-keys")
+
+	/*
+		tests := []struct {
+			name           string
+			fingerprint    string
+			expectedToPass bool
+		}{
+			{
+				name:           "invalid key",
+				fingerprint:    invalidKeyFingerprint,
+				expectedToPass: false,
+			},
+			{
+				name:           "valid key",
+				fingerprint:    validKeyFingerprint,
+				expectedToPass: true,
+			},
+		}
+
+		for _, tt := range tests {
+			t.Run(tt.name, func(t *testing.T) {
+			})
+		}
+	*/
+	err = keyring.ExportPrivateKey(privKeyExportFile, false)
+	if err == nil {
+		t.Fatalf("test successfully loaded an empty keyring but should have failed")
+	}
+
+	err = keyring.ExportPubKey(pubKeyExportFile, false)
+	if err == nil {
+		t.Fatalf("test successfully loaded an empty keyring but should have failed")
+	}
+
+	// Setup the input file that will act as stdin
+	tempFile, err := ioutil.TempFile("", "inputFile-")
+	if err != nil {
+		t.Fatal("failed to create temporary file:", err)
+	}
+	defer tempFile.Close()
+	defer os.Remove(tempFile.Name())
+
+	_, err = tempFile.Write([]byte(validKeyGenInput))
+	if err != nil {
+		t.Fatalf("failed to write to %s: %s", tempFile.Name(), err)
+	}
+
+	// reposition to the beginning of the file to have something to read
+	_, err = tempFile.Seek(0, 0)
+	if err != nil {
+		t.Fatalf("failed to reposition to beginning of file %s: %s", tempFile.Name(), err)
+	}
+
+	// Redirect stdin
 	savedStdin := os.Stdin
 	defer func() {
 		os.Stdin = savedStdin
 	}()
-	os.Stdin = testFile
+	os.Stdin = tempFile
 
-	// Actual test, run the test with the first line
-	output, err := AskQuestion("Question test: ")
+	_, err = keyring.GenKeyPair(myURI, myToken)
 	if err != nil {
-		t.Fatalf("failed to get response from AskQuestion(): %s", err)
+
+		t.Fatalf("failed to create new keys: %s", err)
 	}
 
-	// We analyze the result. We always make sure we do not get the '\n'
-	firstAnswer := testStr[:strings.Index(testStr, "\n")]
-	restAnswer := testStr[len(firstAnswer)+1:]
-	if output != firstAnswer {
-		t.Fatal("AskQuestion() returned", output, "instead of", firstAnswer)
-	}
-
-	// Test with the second line
-	output, err = AskQuestion("Question test 2: ")
+	_, err = tempFile.Seek(0, 0)
 	if err != nil {
-		t.Fatalf("failed to get response: %s", err)
+		t.Fatalf("failed to reposition to beginning of file %s: %s", tempFile.Name(), err)
 	}
-
-	// We analyze the result
-	secondAnswer := restAnswer[:strings.Index(restAnswer, "\n")]
-	if output != secondAnswer {
-		t.Fatalf("AskQuestion() returned: %s instead of: %s", output, secondAnswer)
-	}
-
-	// Test with the third line
-	output, err = AskQuestion("Question test 3: ")
+	keySelectionInput := "0\n" + dummyPassPhrase
+	_, err = tempFile.Write([]byte(keySelectionInput))
 	if err != nil {
-		t.Fatalf("failed to get response: %s", err)
+		t.Fatalf("failed to write to %s: %s", tempFile.Name(), err)
 	}
 
-	// We analyze the result
-	if output != "" {
-		t.Fatalf("AskQuestion() returned: %s instead of an empty string", output)
-	}
-
-	// Test with the final line
-	output, err = AskQuestion("Question test 4: ")
+	// reposition to the beginning of the file to have something to read
+	_, err = tempFile.Seek(0, 0)
 	if err != nil {
-		t.Fatalf("failed to get response: %s", err)
+		t.Fatalf("failed to reposition to beginning of file %s: %s", tempFile.Name(), err)
 	}
 
-	finalAnswer := restAnswer[len(secondAnswer)+2:] // We have to account for 2 "\n"
-	if output != finalAnswer {
-		t.Fatalf("AskQuestion() returned: %s instead of: %s", output, finalAnswer)
-	}
-}
-
-func TestAskQuestionNoEcho(t *testing.T) {
-	test.DropPrivilege(t)
-	defer test.ResetPrivilege(t)
-
-	testStr := "test test\ntest2 test2 test2\n\ntest3 test3 test3 test3"
-	testBytes := []byte(testStr)
-
-	// We create a temporary file that will act as stdin
-	testFile, err := ioutil.TempFile("", "inputTest")
+	err = keyring.ExportPrivateKey(privKeyExportFile, false)
 	if err != nil {
-		t.Fatalf("failed to create temporary file: %s", err)
+		t.Fatalf("failed to load the private key: %s", err)
 	}
-	defer testFile.Close()
-	defer os.Remove(testFile.Name())
-
-	// Write the data that AskQuestionNoEcho will later on read
-	_, err = testFile.Write(testBytes)
-	if err != nil {
-		t.Fatalf("failed to write to temporary file: %s", err)
-	}
-
-	// Reposition to the beginning to ensure there is data to read
-	_, err = testFile.Seek(0, 0)
-	if err != nil {
-		t.Fatalf("failed to reposition to beginning of file: %s", err)
-	}
-
-	// Redirect Stdin
-	savedStdin := os.Stdin
-	defer func() {
-		os.Stdin = savedStdin
-	}()
-	os.Stdin = testFile
-
-	// Test AskQuestionNoEcho(), starting with the first line
-	output, err := AskQuestionNoEcho("Test question 1: ")
-	if err != nil {
-		t.Fatalf("failed to get output from AskQuestionNoEcho(): %s", err)
-	}
-
-	// Analyze the result
-	firstAnswer := testStr[:strings.Index(testStr, "\n")]
-	restAnswer := testStr[len(firstAnswer)+1:] // Ignore "\n"
-	if output != firstAnswer {
-		t.Fatalf("AskQuestionNoEcho() returned %s instead of %s", output, firstAnswer)
-	}
-
-	// Test with the second line
-	output, err = AskQuestionNoEcho("Test question 2: ")
-	if err != nil {
-		t.Fatalf("failed to get output from AskQuestionNoEcho(): %s", err)
-	}
-
-	// We analyze the answer
-	secondAnswer := restAnswer[:strings.Index(restAnswer, "\n")]
-	if output != secondAnswer {
-		t.Fatalf("AskQuestionNoEcho() returned %s instead of %s", output, secondAnswer)
-	}
-
-	// Test with third line
-	output, err = AskQuestionNoEcho("Test question 3: ")
-	if err != nil {
-		t.Fatalf("failed to get output from AskQuestionNoEcho(): %s", err)
-	}
-
-	// We analyze the answer
-	if output != "" {
-		t.Fatalf("AskQuestionNoEcho() returned %s instead of an empty string", output)
-	}
-
-	// Test with the final line
-	output, err = AskQuestionNoEcho("Test question 4: ")
-	if err != nil {
-		t.Fatalf("failed to get output from AskQuestionNoEcho(): %s", err)
-	}
-
-	finalAnswer := restAnswer[len(secondAnswer)+2:] // We have to account for 2 "\n"
-	if output != finalAnswer {
-		t.Fatalf("AskQuestionNoEcho() returned %s instead of %s", output, finalAnswer)
-	}
+	/*
+		err = keyring.ExportPubKey(pubKeyExportFile, false)
+		if err != nil {
+			t.Fatalf("failed to load the public key: %s", err)
+		}
+	*/
 }
 
 func TestGenKeyPair(t *testing.T) {
 	test.DropPrivilege(t)
 	defer test.ResetPrivilege(t)
-
-	myToken := "MyToken"
-	myURI := "MyURI"
 
 	// Prepare all the answers that GenKeyPair() is expecting.
 	tests := []struct {
@@ -696,7 +653,7 @@ func TestGenKeyPair(t *testing.T) {
 	}{
 		{
 			name:      "valid case",
-			input:     "A tester\ntest@my.info\n\nfakepassphrase\nfakepassphrase\nn\n",
+			input:     validKeyGenInput,
 			shallPass: true,
 		},
 		{
@@ -745,66 +702,6 @@ func TestGenKeyPair(t *testing.T) {
 			_, err = keyring.GenKeyPair(myURI, myToken)
 			if tt.shallPass && err != nil {
 				t.Fatalf("valid case %s failed: %s", tt.name, err)
-			}
-			if !tt.shallPass && err == nil {
-				t.Fatalf("invalid case %s succeeded", tt.name)
-			}
-		})
-	}
-}
-
-func TestGetPassphrase(t *testing.T) {
-	test.DropPrivilege(t)
-	defer test.ResetPrivilege(t)
-
-	tests := []struct {
-		name      string
-		input     string
-		shallPass bool
-	}{
-		{
-			name:      "valid case",
-			input:     "mypassphrase\nmypassphrase\n",
-			shallPass: true,
-		},
-		{
-			name:      "unmatching passphrases",
-			input:     "mypassphrase\nsomethingelse\n",
-			shallPass: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create a temporary file that will act as input from stdin
-			tempFile, err := ioutil.TempFile("", "inputFile-")
-			if err != nil {
-				t.Fatal("failed to create temporary file:", err)
-			}
-			defer tempFile.Close()
-			defer os.Remove(tempFile.Name())
-
-			// Populate the file
-			_, err = tempFile.Write([]byte(tt.input))
-			if err != nil {
-				t.Fatalf("failed to write data to %s: %s", tempFile.Name(), err)
-			}
-
-			// Re-position to the beginning of file to have something to read
-			_, err = tempFile.Seek(0, 0)
-			if err != nil {
-				t.Fatalf("failed to seek to beginning of %s: %s", tempFile.Name(), err)
-			}
-
-			// Redirect stdin
-			savedStdin := os.Stdin
-			defer func() {
-				os.Stdin = savedStdin
-			}()
-			os.Stdin = tempFile
-
-			pass, err := GetPassphrase("Test: ", 1)
-			if tt.shallPass && (err != nil || pass != "mypassphrase") {
-				t.Fatalf("test %s is expected to succeed but failed: %s", tt.name, err)
 			}
 			if !tt.shallPass && err == nil {
 				t.Fatalf("invalid case %s succeeded", tt.name)
